fixed delete change request when no reference to selected request type is created

+++ spp_change_request/models/change_request.py
@@ -200,7 +200,8 @@ class ChangeRequestBase(models.Model):
             # Only allow the deletion of draft change requests by the user who created it
             if rec.state == "draft" and rec.create_uid == self.env.user:
                 # Remove the associated CR type record
-                rec.request_type_ref_id.unlink()
+                if rec.request_type_ref_id:
+                    rec.request_type_ref_id.unlink()
                 return super(ChangeRequestBase, self).unlink()
             else:
                 raise UserError(
